requirements.txt
fastapi==0.120.1
uvicorn==0.23.2
pydantic==2.8.0
httpx==0.28.1
sqlalchemy==2.0.44
asyncpg==0.30.0
python-dotenv==1.0.1
alembic==1.11.1
pytest==8.4.2
python-jose[cryptography]==3.5.0
pydantic_settings==2.11.0
bcrypt==5.0.0